backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
PyPDF2
reportlab
python-multipart
